airsim/train/requirements.txt
numpy
colorama
matplotlib
scipy
tensorboard
tensorboardx
tensorflow
pytorch
